scaling DFS FF. still have to fix timestamp reachable hyperedges again

+++ datastructure/reachable_checker.h
@@ -177,7 +177,7 @@ namespace whfc {
 	
 	private:
 		BitsetReachableHyperedges bits;
-		TimestampReachableHyperedges<uint8_t> timestamps;
+		TimestampReachableHyperedges<uint32_t> timestamps;
 	};
